+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+ - provide ETH values as well as validator numbers in "epoch summary"
+
 1.37.3:
  - add "hoodi" to the list of supported networks